refactor(cli): share component filter/dependency logic, drop sibling-component expansion (DX-273)
